[TCcnt] Changed TCcountRets to size_t and fixed most warnings

+++ src/libsac2c/tree/tree_compound.h
@@ -423,7 +423,7 @@ extern node *TCgetNthArg (int n, node *args);
  ***  N_ret :
  ***/
 
-extern int TCcountRets (node *rets);
+extern size_t TCcountRets (node *rets);
 extern int TCcountRetsIgnoreArtificials (node *rets);
 extern node *TCappendRet (node *chain, node *item);
 extern node *TCcreateIdsFromRets (node *rets, node **vardecs);
