+++ src/libsac2c/types/types.h
@@ -639,7 +639,7 @@ typedef enum {
 } te_kind_t;
 
 typedef struct TE_INFO te_info;
-typedef int (*te_funptr) (ntype *args);
+typedef size_t (*te_funptr) (ntype *args);
 
 /*******************************************************************************
  *
